Modules
- added a module with all the JS behavior following ES6 good practices
